Refactored encodeState/decodeState into AuthUtils
These are used to encode the private resource URL that should
be redirected back to. This allows us to know what WApplication
is doing the authorization/authentication, and should even work
when Wt is in dedicated process mode as long as the secret is the
same for parent and child process.